cosmetic changes

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: smelicha <smelicha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 16:06:00 by smelicha          #+#    #+#             */
-/*   Updated: 2023/03/28 17:50:24 by smelicha         ###   ########.fr       */
+/*   Updated: 2023/04/17 23:52:35 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
